do a heckin propagation
