+++ angr/analyses/forward_analysis/visitors/call_graph.py
@@ -1,4 +1,4 @@
-from ...cfg.cfg_utils import CFGUtils
+from angr.utils.graph import GraphUtils
 from .graph import GraphVisitor
 
 
@@ -20,7 +20,7 @@ class CallGraphVisitor(GraphVisitor):
         return list(self.callgraph.predecessors(node))
 
     def sort_nodes(self, nodes=None):
-        sorted_nodes = CFGUtils.quasi_topological_sort_nodes(self.callgraph)
+        sorted_nodes = GraphUtils.quasi_topological_sort_nodes(self.callgraph)
 
         if nodes is not None:
             sorted_nodes = [n for n in sorted_nodes if n in set(nodes)]
