+++ angr/analyses/xrefs.py
@@ -10,7 +10,7 @@ from .propagator.vex_vars import VEXTmp
 from .propagator.values import Top
 from . import register_analysis, PropagatorAnalysis
 from .analysis import Analysis
-from .forward_analysis import FunctionGraphVisitor, SingleNodeGraphVisitor, ForwardAnalysis
+from angr.analyses import visitors, ForwardAnalysis
 
 
 class SimEngineXRefsVEX(
@@ -186,13 +186,13 @@ class XRefsAnalysis(ForwardAnalysis, Analysis):  # pylint:disable=abstract-metho
             if block is not None:
                 raise ValueError('You cannot specify both "func" and "block".')
             # traversing a function
-            graph_visitor = FunctionGraphVisitor(func, func_graph)
+            graph_visitor = visitors.FunctionGraphVisitor(func, func_graph)
             if replacements is None:
                 prop = self.project.analyses[PropagatorAnalysis].prep()(func=func, func_graph=func_graph)
                 replacements = prop.model.replacements
         elif block is not None:
             # traversing a block
-            graph_visitor = SingleNodeGraphVisitor(block)
+            graph_visitor = visitors.SingleNodeGraphVisitor(block)
             if replacements is None:
                 prop = self.project.analyses[PropagatorAnalysis].prep()(block=block)
                 replacements = prop.model.replacements
