+++ angr/knowledge_plugins/propagations/__init__.py
@@ -1,2 +1,2 @@
-from .propagation_manager import PropagationManager
 from .propagation_model import PropagationModel
+from .propagation_manager import PropagationManager
