+++ angr/knowledge_plugins/propagations/propagation_model.py
@@ -1,7 +1,11 @@
 from typing import Tuple, Dict, Set, DefaultDict, Any, Optional
 from collections import defaultdict
 
+import claripy
+import ailment
 from angr.serializable import Serializable
+from angr.knowledge_plugins.functions.function import Function
+from .states import PropagatorVEXState, PropagatorAILState, PropagatorState
 
 
 class PropagationModel(Serializable):
@@ -18,6 +22,8 @@ class PropagationModel(Serializable):
         "equivalence",
         # internals of the function graph visitor
         "graph_visitor",
+        "_initial_state",
+        "_function",
     )
 
     def __init__(
@@ -28,6 +34,7 @@ class PropagationModel(Serializable):
         block_initial_reg_values: Optional[Dict] = None,
         replacements: Optional[DefaultDict[Any, Dict]] = None,
         equivalence: Optional[Set] = None,
+        function: Optional[Function] = None,
     ):
         self.key = prop_key
         self.node_iterations = node_iterations if node_iterations is not None else defaultdict(int)
@@ -37,9 +44,29 @@ class PropagationModel(Serializable):
         self.equivalence = equivalence if equivalence is not None else set()
 
         self.graph_visitor = None
+        self._initial_state = None
+        self._function = function
 
     def downsize(self):
         self.node_iterations = None
         self.block_initial_reg_values = None
         self.states = None
         self.graph_visitor = None
+
+    def block_beginning_state(self, block_addr) -> PropagatorState:
+        if self._function is None:
+            raise NotImplementedError
+        else:
+            node = self._function.get_node(block_addr)
+            preds = [self.states[pnode.addr] for pnode in self._function.graph.predecessors(node)]
+            if not preds:
+                if isinstance(node, ailment.Block):
+                    state = PropagatorAILState.initial_state(self._function.project, func_addr=self._function.addr)
+                else:
+                    state = PropagatorVEXState.initial_state(self._function.project, func_addr=self._function.addr)
+                    state.store_register(
+                        state.arch.ip_offset, state.arch.bytes, claripy.BVV(block_addr, state.arch.bits)
+                    )
+            else:
+                state, _ = preds[0].merge(*preds[1:])
+            return state
